+++ src/deion/worker/api/index.ts
@@ -53,6 +53,8 @@ import type {
 	TradeTeam,
 	Options,
 	ExpansionDraftSetupTeam,
+	RealTeamInfo,
+	RealPlayerPhotos,
 } from "../../common/types";
 import setGameAttributes from "../core/league/setGameAttributes";
 
@@ -215,6 +217,33 @@ const createLeague = async (
 	difficulty: number,
 	importLid: number | undefined | null,
 ): Promise<number> => {
+	if (leagueFile.players) {
+		const realPlayerPhotos = (await idb.meta.get(
+			"attributes",
+			"realPlayerPhotos",
+		)) as RealPlayerPhotos | undefined;
+		if (realPlayerPhotos) {
+			for (const p of leagueFile.players) {
+				if (p.srID && realPlayerPhotos[p.srID]) {
+					p.imgURL = realPlayerPhotos[p.srID];
+				}
+			}
+		}
+	}
+
+	if (leagueFile.teams) {
+		const realTeamInfo = (await idb.meta.get("attributes", "realTeamInfo")) as
+			| RealTeamInfo
+			| undefined;
+		if (realTeamInfo) {
+			for (const t of leagueFile.teams) {
+				if (t.srID && realTeamInfo[t.srID]) {
+					Object.assign(t, realTeamInfo[t.srID]);
+				}
+			}
+		}
+	}
+
 	const lid = league.create({
 		name,
 		tid,
@@ -1553,8 +1582,97 @@ const updateMultiTeamMode = async (gameAttributes: {
 	await toUI("realtimeUpdate", [["g.userTids"]]);
 };
 
-const updateOptions = async (options: Options) => {
+const updateOptions = async (
+	options: Options & {
+		realPlayerPhotos: string;
+		realTeamInfo: string;
+	},
+) => {
+	let realPlayerPhotos;
+	let realTeamInfo;
+	if (options.realPlayerPhotos !== "") {
+		try {
+			realPlayerPhotos = JSON.parse(options.realPlayerPhotos);
+		} catch (err) {
+			console.log(err);
+			throw new Error("Invalid JSON in real player photos");
+		}
+		if (typeof realPlayerPhotos !== "object") {
+			throw new Error(
+				"Invalid data format in real player photos - input is not an object",
+			);
+		}
+		for (const [key, value] of Object.entries(realPlayerPhotos)) {
+			if (typeof value !== "string") {
+				throw new Error(
+					`Invalid data format in real player photos - value for "${key}" is not a string`,
+				);
+			}
+		}
+	}
+	if (options.realTeamInfo !== "") {
+		try {
+			realTeamInfo = JSON.parse(options.realTeamInfo);
+		} catch (err) {
+			console.log(err);
+			throw new Error("Invalid JSON in real team info");
+		}
+		if (typeof realTeamInfo !== "object") {
+			throw new Error(
+				"Invalid data format in real team info - input is not an object",
+			);
+		}
+		const strings = ["abbrev", "region", "name", "imgURL"];
+		const numbers = ["pop"];
+		for (const [abbrev, teamInfo] of Object.entries(realTeamInfo)) {
+			if (typeof teamInfo !== "object") {
+				throw new Error(
+					"Invalid data format in real team info - input is not an object",
+				);
+			}
+			for (const [key, value] of Object.entries(teamInfo as any)) {
+				if (strings.includes(key)) {
+					if (typeof value !== "string") {
+						throw new Error(
+							`Invalid data format in real team info - value for "${abbrev}.${key}" is not a string`,
+						);
+					}
+				} else if (numbers.includes(key)) {
+					if (typeof value !== "number") {
+						throw new Error(
+							`Invalid data format in real team info - value for "${abbrev}.${key}" is not a number`,
+						);
+					}
+				} else if (key === "colors") {
+					if (!Array.isArray(value)) {
+						throw new Error(
+							`Invalid data format in real team info - value for "${abbrev}.${key}" is not an array`,
+						);
+					}
+					if (value.length !== 3) {
+						throw new Error(
+							`Invalid data format in real team info - value for "${abbrev}.${key}" should have 3 colors`,
+						);
+					}
+					for (const color of value) {
+						if (typeof color !== "string") {
+							throw new Error(
+								`Invalid data format in real team info - value for "${abbrev}.${key}" is not an array of strings`,
+							);
+						}
+					}
+				} else {
+					throw new Error(
+						`Invalid data format in real team info - unknown property "${abbrev}.${key}"`,
+					);
+				}
+			}
+		}
+	}
+
 	await idb.meta.put("attributes", options, "options");
+	await idb.meta.put("attributes", realPlayerPhotos, "realPlayerPhotos");
+	await idb.meta.put("attributes", realTeamInfo, "realTeamInfo");
 	await toUI("updateLocal", [{ units: options.units }]);
 	await toUI("realtimeUpdate", [["options"]]);
 };