+++ src/deion/worker/db/reset.ts
@@ -1,6 +1,5 @@
 import { league } from "../core";
 import { idb } from ".";
-import { deleteDB } from "idb";
 import { logEvent } from "../util";
 
 const reset = async () => {
@@ -12,22 +11,16 @@ const reset = async () => {
 		leagues.map(async l => {
 			await league.remove(l.lid);
 			numDeleted += 1;
-			console.log(`Deleted ${numDeleted} of ${leagues.length} leagues`);
+			logEvent({
+				type: "info",
+				text: `Deleted ${numDeleted} of ${leagues.length} leagues...`,
+				saveToDb: false,
+			});
 		}),
 	);
 
-	// Delete any current meta database
-	console.log("Deleting any current meta database...");
-	idb.meta.close();
-	await deleteDB("meta", {
-		blocked() {
-			logEvent({
-				type: "error",
-				text: "Please close any other open tabs.",
-				saveToDb: false,
-			});
-		},
-	});
+	// Delete leagues from meta database
+	await idb.meta.transaction("leagues", "readwrite").store.clear();
 };
 
 export default reset;